Collapse common spec fields into a CommonSpec.
